+++ app/charts/column/chart-column.tsx
@@ -39,7 +39,7 @@ export const ChartColumnsVisualization = (
 };
 
 const ChartColumns = memo((props: ChartProps<ColumnConfig>) => {
-  const { chartConfig, dimensions } = props;
+  const { chartConfig, dimensions, dimensionsById } = props;
   const { fields, interactiveFiltersConfig } = chartConfig;
   const filters = useChartConfigFilters(chartConfig);
   const [{ dashboardFilters }] = useConfiguratorState(hasChartConfigs);
@@ -70,11 +70,13 @@ const ChartColumns = memo((props: ChartProps<ColumnConfig>) => {
               />
             )}
             <LegendColor
+              dimensionsById={dimensionsById}
               chartConfig={chartConfig}
               symbol="square"
               interactive={
                 fields.segment && interactiveFiltersConfig?.legend.active
               }
+              showTitle={fields.segment && fields.segment.showTitle}
             />
           </ChartControlsContainer>
         </StackedColumnsChart>
@@ -101,11 +103,13 @@ const ChartColumns = memo((props: ChartProps<ColumnConfig>) => {
               />
             )}
             <LegendColor
+              dimensionsById={dimensionsById}
               chartConfig={chartConfig}
               symbol="square"
               interactive={
                 fields.segment && interactiveFiltersConfig?.legend.active
               }
+              showTitle={fields.segment && fields.segment.showTitle}
             />
           </ChartControlsContainer>
         </GroupedColumnChart>