+++ app/configurator/components/filters.tsx
@@ -106,6 +106,7 @@ import useEvent from "@/utils/use-event";
 
 import { interlace } from "../../utils/interlace";
 import { getTimeFilterOptions } from "../../utils/time-filter-options";
+import { useLegendTitleVisibility } from "../configurator-state/segment-config-state";
 
 const useStyles = makeStyles((theme: Theme) => {
   return {
@@ -354,44 +355,82 @@ const MultiFilterContent = ({
   });
 
   const interactiveFilterProps = useInteractiveFiltersToggle("legend");
+  const visibleLegendProps = useLegendTitleVisibility();
   const chartSymbol = getChartSymbol(chartConfig.chartType);
 
   return (
     <Box sx={{ position: "relative" }}>
       <Box mb={4}>
-        <Box sx={{ justifyContent: "space-between", display: "flex" }}>
+        <Box
+          sx={{
+            justifyContent: "space-between",
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "flex-end",
+          }}
+        >
           {chartConfig.activeField === "segment" ? (
-            <FormControlLabel
-              componentsProps={{ typography: { variant: "body2" } }}
-              control={<Switch {...interactiveFilterProps} />}
-              label={
-                <MaybeTooltip
-                  tooltipProps={{ enterDelay: 600 }}
-                  title={
-                    <TooltipTitle
-                      text={
-                        <Trans id="controls.filters.interactive.tooltip">
-                          Allow users to change filters
-                        </Trans>
+            <>
+              <Flex sx={{ py: 1, alignItems: "center", width: "100%" }}>
+                <FormControlLabel
+                  componentsProps={{ typography: { variant: "body2" } }}
+                  control={<Switch {...interactiveFilterProps} />}
+                  label={
+                    <MaybeTooltip
+                      tooltipProps={{ enterDelay: 600 }}
+                      title={
+                        <TooltipTitle
+                          text={
+                            <Trans id="controls.filters.interactive.tooltip">
+                              Allow users to change filters
+                            </Trans>
+                          }
+                        />
                       }
-                    />
+                    >
+                      <div>
+                        <Trans id="controls.filters.interactive.toggle">
+                          Interactive
+                        </Trans>
+                      </div>
+                    </MaybeTooltip>
                   }
-                >
-                  <div>
-                    <Trans id="controls.filters.interactive.toggle">
-                      Interactive
-                    </Trans>
-                  </div>
-                </MaybeTooltip>
-              }
-            />
+                />
+              </Flex>
+              <Flex sx={{ py: 1, alignItems: "center", width: "100%" }}>
+                <FormControlLabel
+                  componentsProps={{ typography: { variant: "body2" } }}
+                  control={<Switch {...visibleLegendProps} />}
+                  label={
+                    <MaybeTooltip
+                      tooltipProps={{ enterDelay: 600 }}
+                      title={
+                        <TooltipTitle
+                          text={
+                            <Trans id="controls.filters.show-legend-toggle">
+                              Allow users to change Legend visibility
+                            </Trans>
+                          }
+                        />
+                      }
+                    >
+                      <div>
+                        <Trans id="controls.filters.show-legend.toggle">
+                          Show legend titles
+                        </Trans>
+                      </div>
+                    </MaybeTooltip>
+                  }
+                />
+              </Flex>
+            </>
           ) : null}
           <Button
             variant="contained"
             size="small"
             color="primary"
             onClick={handleOpenAutocomplete}
-            sx={{ justifyContent: "center", mb: 2 }}
+            sx={{ justifyContent: "center", mb: 2, width: "fit-content" }}
           >
             <Trans id="controls.set-filters">Edit filters</Trans>
           </Button>
@@ -940,9 +979,9 @@ export const DimensionValuesMultiFilter = ({
   const chartConfig = getChartConfig(state);
   const getValueColor = useEvent((value: string) => {
     const colorPath = getPathToColorConfigProperty({
-      field,
+      field: isColorInConfig(chartConfig) ? "color" : field,
       colorConfigPath,
-      propertyPath: `colorMapping["${value}"]`,
+      propertyPath: `["${value}"]`,
     });
     return get(chartConfig, colorPath);
   });