+++ app/configurator/configurator-state/reducer.tsx
@@ -41,10 +41,10 @@ import {
   DashboardTimeRangeFilter,
   enableLayouting,
   Filters,
-  GenericField,
   GenericFields,
   isAreaConfig,
   isColorInConfig,
+  isSegmentInConfig,
   isTableConfig,
   ReactGridLayoutType,
 } from "@/config-types";
@@ -429,7 +429,7 @@ export const handleChartFieldChanged = (
 
 export const handleChartOptionChanged = (
   draft: ConfiguratorState,
-  action: Extract<ConfiguratorStateAction, { type: "COLOR_MAPPING_UPDATED" }>
+  action: Extract<ConfiguratorStateAction, { type: "COLOR_FIELD_UPDATED" }>
 ) => {
   if (isConfiguring(draft)) {
     const { locale, path, field, value } = action.value;
@@ -478,7 +478,6 @@ export const updateColorMapping = (
       field,
       colorConfigPath,
       colorMapping: oldColorMapping,
-      dimensionId,
       values,
       random,
     } = action.value;
@@ -503,17 +502,12 @@ export const updateColorMapping = (
         });
       }
     } else {
-      const fieldValue: (GenericField & { paletteId: string }) | undefined =
-        get(chartConfig, path);
-
-      if (fieldValue?.componentId === dimensionId) {
-        colorMapping = mapValueIrisToColor({
-          paletteId: fieldValue.paletteId,
-          dimensionValues: values,
-          colorMapping: oldColorMapping,
-          random,
-        });
-      }
+      colorMapping = mapValueIrisToColor({
+        paletteId: "dimension",
+        dimensionValues: values,
+        colorMapping: oldColorMapping,
+        random,
+      });
     }
 
     if (colorMapping) {
@@ -585,7 +579,7 @@ const handleAddNewChartConfig = (
   draft: ConfiguratorState,
   chartConfig: ChartConfig
 ) => {
-  if (isConfiguring(draft)) {
+  if (isConfiguring(draft) || isLayouting(draft)) {
     draft.chartConfigs.push(chartConfig);
     draft.activeChartKey = chartConfig.key;
     draft.layout.blocks.push({
@@ -662,6 +656,20 @@ const reducer_: Reducer<ConfiguratorState, ConfiguratorStateAction> = (
     case "CHART_FIELD_CHANGED":
       return handleChartFieldChanged(draft, action);
 
+    case "CHART_SHOW_LEGEND_TITLE_CHANGED":
+      if (isConfiguring(draft)) {
+        const chartConfig = getChartConfig(draft);
+
+        if (isSegmentInConfig(chartConfig) && chartConfig.fields.segment) {
+          chartConfig.fields.segment.showTitle = action.value;
+        }
+        const index = draft.chartConfigs.findIndex(
+          (d) => d.key === chartConfig.key
+        );
+        draft.chartConfigs[index] = chartConfig;
+      }
+      return draft;
+
     case "CHART_FIELD_DELETED":
       if (isConfiguring(draft)) {
         const chartConfig = getChartConfig(draft);
@@ -698,7 +706,7 @@ const reducer_: Reducer<ConfiguratorState, ConfiguratorStateAction> = (
 
       return draft;
 
-    case "COLOR_MAPPING_UPDATED":
+    case "COLOR_FIELD_UPDATED":
       return handleChartOptionChanged(draft, action);
 
     case "COLOR_FIELD_SET":