+++ app/login/components/profile-content-tabs.tsx
@@ -11,11 +11,13 @@ import { ProfileVisualizationsTable } from "@/login/components/profile-tables";
 import { useRootStyles } from "@/login/utils";
 import useEvent from "@/utils/use-event";
 
+import { ProfileColorPaletteContent } from "./color-palettes/profile-color-palette-content";
+
 const useStyles = makeStyles<Theme>((theme) => ({
   section: {
     borderBottom: 1,
     borderColor: "divider",
-    marginTop: theme.spacing(6),
+    paddingTop: theme.spacing(6),
   },
   tabList: {
     minHeight: "fit-content",
@@ -83,8 +85,16 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
               <Tab
                 className={classes.tab}
                 label={t({
+                  id: "login.profile.my-drafts",
+                  message: "My Drafts",
+                })}
+                value="drafts"
+              />
+              <Tab
+                className={classes.tab}
+                label={t({
                   id: "login.profile.my-published-visualizations",
-                  message: "My published visualizations",
+                  message: "My Visualizations",
                 })}
                 value="published"
               />
@@ -92,10 +102,10 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
                 data-testId="color-palettes-tab"
                 className={classes.tab}
                 label={t({
-                  id: "login.profile.my-drafts",
-                  message: "My draft visualizations",
+                  id: "login.profile.my-color-palettes",
+                  message: "My Color Palettes",
                 })}
-                value="drafts"
+                value="palettes"
               />
             </TabList>
           </Box>
@@ -109,7 +119,7 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
           <ProfileVisualizationsTable
             title={t({
               id: "login.profile.my-drafts",
-              message: "My draft visualizations",
+              message: "My Drafts",
             })}
             userId={userId}
             userConfigs={draftConfigs}
@@ -119,7 +129,7 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
           <ProfileVisualizationsTable
             title={t({
               id: "login.profile.my-published-visualizations",
-              message: "My published visualizations",
+              message: "My Visualizations",
             })}
             userId={userId}
             userConfigs={publishedConfigs}
@@ -133,7 +143,7 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
           <ProfileVisualizationsTable
             title={t({
               id: "login.profile.my-published-visualizations",
-              message: "My published visualizations",
+              message: "My Visualizations",
             })}
             userId={userId}
             userConfigs={publishedConfigs}
@@ -145,10 +155,21 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
           <ProfileVisualizationsTable
             title={t({
               id: "login.profile.my-drafts",
-              message: "My draft visualizations",
+              message: "My Drafts",
             })}
             userId={userId}
             userConfigs={draftConfigs}
+          />
+        </Box>
+      </TabPanel>
+      <TabPanel className={classes.tabPanel} value="palettes">
+        <Box className={classes.tabPanelContent}>
+          <ProfileColorPaletteContent
+            title={t({
+              id: "login.profile.my-color-palettes",
+              message: "My Color Palettes",
+            })}
+            userId={userId}
           />
         </Box>
       </TabPanel>