+++ app/pages/api/user/color-palette.ts
@@ -1,0 +1,11 @@
+import { api } from "@/server/nextkit";
+import UserController from "@/server/user-controller";
+
+const route = api({
+  POST: UserController.createPalette,
+  GET: UserController.getPalettes,
+  DELETE: UserController.deletePalette,
+  PUT: UserController.updatePalette,
+});
+
+export default route;